requirements.txt
netaddr==0.8.0
pandas==1.3.5
psutil==5.9.2
pyfiglet==0.8.post0
PyYAML==6.0
rich==12.6.0
tabulate==0.8.9
termcolor==2.1.0
user_agents==2.2.0
